refactor(auto-dent): share work-cycle progress rendering (#1257)
